Implement redux and react-router and fixed the issue of Ability to specify SVG canvas width and height #9
